Allow fractional scroll in raw mouse events

+++ src/platform_impl/windows/event_loop.rs
@@ -2134,11 +2134,12 @@ unsafe fn handle_raw_input<T: 'static>(
         }
 
         if util::has_flag(mouse.usButtonFlags, winuser::RI_MOUSE_WHEEL) {
-            let delta = mouse.usButtonData as SHORT / winuser::WHEEL_DELTA;
+            // We must cast to SHORT first, becaues `usButtonData` must be interpreted as signed.
+            let delta = mouse.usButtonData as SHORT as f32 / winuser::WHEEL_DELTA as f32;
             subclass_input.send_event(Event::DeviceEvent {
                 device_id,
                 event: MouseWheel {
-                    delta: LineDelta(0.0, delta as f32),
+                    delta: LineDelta(0.0, delta),
                 },
             });
         }
